:warning: Browser tests aren't working
The browser tests use SauceLabs to run on
[Travis](https://docs.travis-ci.com/user/sauce-connect/).
Forks do not have access to the account in the origin repo therefore
tests aren't working and are timing out (unable to connect).
Rather than have each test take the time and resources to try and run
I've commented them out - the intention being to get an account to run
the tests at some point in the future.